+++ src/main/java/ravo/ravobackend/coldstandby/DatabaseInfo.java
@@ -1,4 +1,4 @@
-package ravo.ravobackend.coldstandby.tasklet;
+package ravo.ravobackend.coldstandby;
 
 import lombok.Getter;
